Added custom hooks in the Popular Flashcard Page
