+++ src/pages/PopularFlashcards/hooks/useFlip.tsx
@@ -1,0 +1,11 @@
+import { useState } from "react";
+
+export const useFlip = () => {
+  const [isFlipped, setIsFlipped] = useState(false);
+
+  const handleFlip = () => {
+    setIsFlipped(!isFlipped);
+  };
+
+  return { isFlipped, handleFlip };
+};
